refactor: Move admin specific cards to a more appropriate place
